pkg/provider: split into pkg/server, pkg/identity
