+++ cmd/captplanet/run.go
@@ -28,8 +28,8 @@ import (
 	"storj.io/storj/pkg/piecestore/psserver"
 	"storj.io/storj/pkg/pointerdb"
 	"storj.io/storj/pkg/process"
-	"storj.io/storj/pkg/provider"
 	"storj.io/storj/pkg/satellite/satelliteweb"
+	"storj.io/storj/pkg/server"
 	"storj.io/storj/pkg/utils"
 	"storj.io/storj/satellite/satellitedb"
 )
@@ -40,7 +40,7 @@ const (
 
 // Satellite is for configuring client
 type Satellite struct {
-	Identity    provider.IdentityConfig
+	Server      server.ServerConfig
 	Kademlia    kademlia.Config
 	PointerDB   pointerdb.Config
 	Overlay     overlay.Config
@@ -58,7 +58,7 @@ type Satellite struct {
 
 // StorageNode is for configuring storage nodes
 type StorageNode struct {
-	Identity provider.IdentityConfig
+	Server   server.ServerConfig
 	Kademlia kademlia.Config
 	Storage  psserver.Config
 }
@@ -66,7 +66,7 @@ type StorageNode struct {
 var (
 	runCmd = &cobra.Command{
 		Use:   "run",
-		Short: "Run all providers",
+		Short: "Run all servers",
 		RunE:  cmdRun,
 	}
 
@@ -100,14 +100,14 @@ func cmdRun(cmd *cobra.Command, args []string) (err error) {
 	// start satellite
 	go func() {
 		_, _ = fmt.Printf("Starting satellite on %s\n",
-			runCfg.Satellite.Identity.Server.Address)
+			runCfg.Satellite.Server.Address)
 
 		if runCfg.Satellite.Audit.SatelliteAddr == "" {
-			runCfg.Satellite.Audit.SatelliteAddr = runCfg.Satellite.Identity.Server.Address
+			runCfg.Satellite.Audit.SatelliteAddr = runCfg.Satellite.Server.Address
 		}
 
 		if runCfg.Satellite.Web.SatelliteAddr == "" {
-			runCfg.Satellite.Web.SatelliteAddr = runCfg.Satellite.Identity.Server.Address
+			runCfg.Satellite.Web.SatelliteAddr = runCfg.Satellite.Server.Address
 		}
 
 		database, err := satellitedb.New(runCfg.Satellite.Database)
@@ -126,7 +126,7 @@ func cmdRun(cmd *cobra.Command, args []string) (err error) {
 		ctx = context.WithValue(ctx, "masterdb", database)
 
 		// Run satellite
-		errch <- runCfg.Satellite.Identity.Run(ctx,
+		errch <- runCfg.Satellite.Server.Run(ctx,
 			grpcauth.NewAPIKeyInterceptor(),
 			runCfg.Satellite.Kademlia,
 			runCfg.Satellite.Audit,
@@ -152,23 +152,23 @@ func cmdRun(cmd *cobra.Command, args []string) (err error) {
 	// start the storagenodes
 	for i, v := range runCfg.StorageNodes {
 		go func(i int, v StorageNode) {
-			identity, err := v.Identity.Load()
+			identity, err := v.Server.Identity.Load()
 			if err != nil {
 				return
 			}
 
-			address := v.Identity.Server.Address
+			address := v.Server.Address
 			storagenode := fmt.Sprintf("%s:%s", identity.ID.String(), address)
 
 			_, _ = fmt.Printf("Starting storage node %d %s (kad on %s)\n", i, storagenode, address)
-			errch <- v.Identity.Run(ctx, nil, v.Kademlia, v.Storage)
+			errch <- v.Server.Run(ctx, nil, v.Kademlia, v.Storage)
 		}(i, v)
 	}
 
 	// start s3 uplink
 	go func() {
 		_, _ = fmt.Printf("Starting s3-gateway on %s\nAccess key: %s\nSecret key: %s\n",
-			runCfg.Uplink.Identity.Server.Address,
+			runCfg.Uplink.Server.Address,
 			runCfg.Uplink.Minio.AccessKey,
 			runCfg.Uplink.Minio.SecretKey)
 		errch <- runCfg.Uplink.Run(ctx)
