+++ pkg/identity/cert_authority_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package provider
+package identity
 
 import (
 	"context"
