+++ pkg/identity/common.go
@@ -1,0 +1,16 @@
+// Copyright (C) 2018 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+package identity
+
+import (
+	"github.com/zeebo/errs"
+	monkit "gopkg.in/spacemonkeygo/monkit.v2"
+)
+
+var (
+	mon = monkit.Package()
+
+	// Error is a pkg/identity error
+	Error = errs.Class("pkg/identity error")
+)
