+++ pkg/identity/identity.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package provider
+package identity
 
 import (
 	"bytes"
@@ -10,11 +10,8 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"io/ioutil"
-	"net"
-	"os"
 
 	"github.com/zeebo/errs"
-	"go.uber.org/zap"
 	"golang.org/x/crypto/sha3"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -59,7 +56,6 @@ type IdentitySetupConfig struct {
 	KeyPath   string `help:"path to the private key for this identity" default:"$CONFDIR/identity.key"`
 	Overwrite bool   `help:"if true, existing identity certs AND keys will overwritten for" default:"false"`
 	Version   string `help:"semantic version of identity storage format" default:"0"`
-	Server    ServerConfig
 }
 
 // IdentityConfig allows you to run a set of Responsibilities with the given
@@ -67,44 +63,12 @@ type IdentitySetupConfig struct {
 type IdentityConfig struct {
 	CertPath string `help:"path to the certificate chain for this identity" default:"$CONFDIR/identity.cert"`
 	KeyPath  string `help:"path to the private key for this identity" default:"$CONFDIR/identity.key"`
-	Server   ServerConfig
-}
-
-// ServerConfig holds server specific configuration parameters
-type ServerConfig struct {
-	RevocationDBURL     string `help:"url for revocation database (e.g. bolt://some.db OR redis://127.0.0.1:6378?db=2&password=abc123)" default:"bolt://$CONFDIR/revocations.db"`
-	PeerCAWhitelistPath string `help:"path to the CA cert whitelist (peer identities must be signed by one these to be verified)"`
-	Address             string `help:"address to listen on" default:":7777"`
-	Extensions          peertls.TLSExtConfig
-}
-
-// ServerOptions holds config, identity, and peer verification function data for use with a grpc server.
-type ServerOptions struct {
-	Config   ServerConfig
-	Ident    *FullIdentity
-	RevDB    *peertls.RevocationDB
-	PCVFuncs []peertls.PeerCertVerificationFunc
-}
-
-// NewServerOptions is a constructor for `serverOptions` given an identity and config
-func NewServerOptions(i *FullIdentity, c ServerConfig) (*ServerOptions, error) {
-	serverOpts := &ServerOptions{
-		Config: c,
-		Ident:  i,
-	}
-
-	err := c.configure(serverOpts)
-	if err != nil {
-		return nil, err
-	}
-
-	return serverOpts, nil
 }
 
 // FullIdentityFromPEM loads a FullIdentity from a certificate chain and
 // private key PEM-encoded bytes
 func FullIdentityFromPEM(chainPEM, keyPEM []byte) (*FullIdentity, error) {
-	chain, err := decodeAndParseChainPEM(chainPEM)
+	chain, err := DecodeAndParseChainPEM(chainPEM)
 	if err != nil {
 		return nil, errs.Wrap(err)
 	}
@@ -280,37 +244,6 @@ func (ic IdentityConfig) Save(fi *FullIdentity) error {
 	)
 }
 
-// Run will run the given responsibilities with the configured identity.
-func (ic IdentityConfig) Run(ctx context.Context, interceptor grpc.UnaryServerInterceptor, responsibilities ...Responsibility) (err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	ident, err := ic.Load()
-	if err != nil {
-		return err
-	}
-
-	lis, err := net.Listen("tcp", ic.Server.Address)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = lis.Close() }()
-
-	opts, err := NewServerOptions(ident, ic.Server)
-	if err != nil {
-		return err
-	}
-	defer func() { err = utils.CombineErrors(err, opts.RevDB.Close()) }()
-
-	s, err := NewProvider(opts, lis, interceptor, responsibilities...)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = s.Close() }()
-
-	zap.S().Infof("Node %s started", s.Identity().ID)
-	return s.Run(ctx)
-}
-
 // RestChainRaw returns the rest (excluding leaf and CA) of the certficate chain as a 2d byte slice
 func (fi *FullIdentity) RestChainRaw() [][]byte {
 	var chain [][]byte
@@ -369,75 +302,6 @@ func (fi *FullIdentity) DialOption(id storj.NodeID) (grpc.DialOption, error) {
 	return grpc.WithTransportCredentials(credentials.NewTLS(tlsConfig)), nil
 }
 
-// NewRevDB returns a new revocation database given the config
-func (c ServerConfig) NewRevDB() (*peertls.RevocationDB, error) {
-	driver, source, err := utils.SplitDBURL(c.RevocationDBURL)
-	if err != nil {
-		return nil, peertls.ErrRevocationDB.Wrap(err)
-	}
-
-	var db *peertls.RevocationDB
-	switch driver {
-	case "bolt":
-		db, err = peertls.NewRevocationDBBolt(source)
-		if err != nil {
-			return nil, peertls.ErrRevocationDB.Wrap(err)
-		}
-		zap.S().Info("Starting overlay cache with BoltDB")
-	case "redis":
-		db, err = peertls.NewRevocationDBRedis(c.RevocationDBURL)
-		if err != nil {
-			return nil, peertls.ErrRevocationDB.Wrap(err)
-		}
-		zap.S().Info("Starting overlay cache with Redis")
-	default:
-		return nil, peertls.ErrRevocationDB.New("database scheme not supported: %s", driver)
-	}
-
-	return db, nil
-}
-
-// configure adds peer certificate verification functions and revocation datase to the config.
-func (c ServerConfig) configure(opts *ServerOptions) (err error) {
-	var pcvs []peertls.PeerCertVerificationFunc
-	parseOpts := peertls.ParseExtOptions{}
-
-	if c.PeerCAWhitelistPath != "" {
-		caWhitelist, err := loadWhitelist(c.PeerCAWhitelistPath)
-		if err != nil {
-			return err
-		}
-		parseOpts.CAWhitelist = caWhitelist
-		pcvs = append(pcvs, peertls.VerifyCAWhitelist(caWhitelist))
-	}
-
-	if c.Extensions.Revocation {
-		opts.RevDB, err = c.NewRevDB()
-		if err != nil {
-			return err
-		}
-		pcvs = append(pcvs, peertls.VerifyUnrevokedChainFunc(opts.RevDB))
-	}
-
-	exts := peertls.ParseExtensions(c.Extensions, parseOpts)
-	pcvs = append(pcvs, exts.VerifyFunc())
-
-	// NB: remove nil elements
-	for i, f := range pcvs {
-		if f == nil {
-			copy(pcvs[i:], pcvs[i+1:])
-			pcvs = pcvs[:len(pcvs)-1]
-		}
-	}
-
-	opts.PCVFuncs = pcvs
-	return nil
-}
-
-func (so *ServerOptions) grpcOpts() (grpc.ServerOption, error) {
-	return so.Ident.ServerOption(so.PCVFuncs...)
-}
-
 func verifyIdentity(id storj.NodeID) peertls.PeerCertVerificationFunc {
 	return func(_ [][]byte, parsedChains [][]*x509.Certificate) error {
 		if id == (storj.NodeID{}) {
@@ -455,20 +319,4 @@ func verifyIdentity(id storj.NodeID) peertls.PeerCertVerificationFunc {
 
 		return nil
 	}
-}
-
-func loadWhitelist(path string) ([]*x509.Certificate, error) {
-	w, err := ioutil.ReadFile(path)
-	if err != nil && !os.IsNotExist(err) {
-		return nil, err
-	}
-
-	var whitelist []*x509.Certificate
-	if w != nil {
-		whitelist, err = decodeAndParseChainPEM(w)
-		if err != nil {
-			return nil, errs.Wrap(err)
-		}
-	}
-	return whitelist, nil
 }
